cmd/smithcmp: add new cmpconn package
This factors out common code needed for connection comparison, which
will be used in an upcoming test.
